backend/requirements.txt
# Core API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0

# Database
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
alembic>=1.14.0

# Git Operations
GitPython>=3.1.43

# GitHub API & OAuth
PyGithub>=2.5.0
httpx>=0.28.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=8.3.0
pytest-cov>=6.0.0

# Utilities
python-multipart>=0.0.20
python-dotenv>=1.0.0
